+++ school_library/models/book.rb
