+++ school_library/models/rental.rb
