+++ school_library/modules/lists.rb
@@ -1,0 +1,23 @@
+require_relative 'create_persons'
+require_relative 'create_books'
+
+module Lists
+  def list_books
+    puts 'List of Books:'
+    Book.all.each do |book|
+      puts "Title: \"#{book.title}\", Author: #{book.author}"
+    end
+  end
+
+  def list_people
+    puts 'List of People (Teachers and Students):'
+
+    Person.all.each do |person|
+      if person.is_a?(Student)
+        puts "[Student] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+      elsif person.is_a?(Teacher)
+        puts "[Teacher] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+      end
+    end
+  end
+end
